Move terminalCommandTracker, node->browser
Part of #69115
