+++ src/tsconfig.strictNullChecks.json
@@ -279,6 +279,7 @@
 		"./vs/workbench/contrib/surveys/electron-browser/nps.contribution.ts",
 
 		"./vs/workbench/contrib/terminal/browser/terminalActions.ts",
+		"./vs/workbench/contrib/terminal/browser/terminalCommandTracker.ts",
 		"./vs/workbench/contrib/terminal/browser/terminalConfigHelper.ts",
 		"./vs/workbench/contrib/terminal/browser/terminalFindWidget.ts",
 		"./vs/workbench/contrib/terminal/browser/terminalLinkHandler.ts",
@@ -295,14 +296,13 @@
 		"./vs/workbench/contrib/terminal/electron-browser/terminalInstance.ts",
 		"./vs/workbench/contrib/terminal/electron-browser/terminalProcessManager.ts",
 		"./vs/workbench/contrib/terminal/node/terminal.ts",
-		"./vs/workbench/contrib/terminal/node/terminalCommandTracker.ts",
 		"./vs/workbench/contrib/terminal/node/terminalProcess.ts",
 		"./vs/workbench/contrib/terminal/node/terminalProcessExtHostProxy.ts",
 		"./vs/workbench/contrib/terminal/node/windowsShellHelper.ts",
 		"./vs/workbench/contrib/terminal/test/electron-browser/terminalColorRegistry.test.ts",
+		"./vs/workbench/contrib/terminal/test/electron-browser/terminalCommandTracker.test.ts",
 		"./vs/workbench/contrib/terminal/test/electron-browser/terminalConfigHelper.test.ts",
 		"./vs/workbench/contrib/terminal/test/electron-browser/terminalLinkHandler.test.ts",
-		"./vs/workbench/contrib/terminal/test/node/terminalCommandTracker.test.ts",
 		"./vs/workbench/contrib/terminal/test/node/terminalEnvironment.test.ts",
 		"./vs/workbench/contrib/themes/browser/themes.contribution.ts",
 		"./vs/workbench/contrib/themes/test/electron-browser/themes.test.contribution.ts",
