+++ src/vs/workbench/contrib/terminal/browser/terminalCommandTracker.ts
